# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator project",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Search and filter articles",
                "description": "Returns a list of articles filtered by keyword, date, categories, sources, and authors",
                "operationId": "getArticles",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search term for article titles or descriptions",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "categories",
                        "in": "query",
                        "description": "Array of category UUIDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "uuid"
                            }
                        }
                    },
                    {
                        "name": "sources",
                        "in": "query",
                        "description": "Array of source UUIDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "uuid"
                            }
                        }
                    },
                    {
                        "name": "authors",
                        "in": "query",
                        "description": "Array of author UUIDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "uuid"
                            }
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Column to order by (e.g. published_at, title)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "published_at"
                        }
                    },
                    {
                        "name": "order_direction",
                        "in": "query",
                        "description": "Order direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of filtered articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get user-specific articles",
                "description": "Returns articles based on the user's saved authors, categories, and sources.",
                "operationId": "getUserArticles",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Column to order by (e.g. published_at, title)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "published_at"
                        }
                    },
                    {
                        "name": "order_direction",
                        "in": "query",
                        "description": "Order direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user-specific articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Article": {
                "title": "Article",
                "description": "Article model representing a news article",
                "required": [
                    "id",
                    "title",
                    "url",
                    "published_at"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the article",
                        "type": "string"
                    },
                    "source_id": {
                        "description": "ID of the source",
                        "type": "string"
                    },
                    "title": {
                        "description": "Article title",
                        "type": "string"
                    },
                    "description": {
                        "description": "Article description",
                        "type": "string"
                    },
                    "content": {
                        "description": "Full content of the article",
                        "type": "string"
                    },
                    "url": {
                        "description": "URL of the article",
                        "type": "string"
                    },
                    "url_to_image": {
                        "description": "URL of the article image",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "Publication date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "source": {
                        "$ref": "#/components/schemas/Source"
                    },
                    "categories": {
                        "description": "Categories of the article",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "authors": {
                        "description": "Authors of the article",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    }
                },
                "type": "object"
            },
            "Author": {
                "title": "Author",
                "description": "Author model representing an author",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the author",
                        "type": "string"
                    },
                    "name": {
                        "description": "Author name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Category model representing a category",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the category",
                        "type": "string"
                    },
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Source": {
                "title": "Source",
                "description": "Source model representing a source",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "UUID of the source",
                        "type": "string"
                    },
                    "external_id": {
                        "description": "Source external id",
                        "type": "string"
                    },
                    "name": {
                        "description": "Source name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}